"Updated data type of 'hs_de' and 'hs_cn' columns in product_groups table to bigInteger, and corrected toastr error messages in product_group_entry.vue"

+++ database/migrations/2024_02_09_121556_create_product_groups_table.php
@@ -16,8 +16,8 @@ return new class extends Migration {
             $table->string('description')->nullable();
             $table->integer('profit')->nullable();
             $table->decimal('amount', 10, 2)->nullable();
-            $table->integer('hs_de');
-            $table->integer('hs_cn');
+            $table->bigInteger('hs_de');
+            $table->bigInteger('hs_cn');
             $table->timestamps();
         });
     }
